support IDEA EAP 2024.1 - adapt to new velocity initialization mechanism
